Simplify server starting/stopping
Remove the custom command orchestra:serve which requires Laravel to be
booted. Ideally we don't want to starting/stopping servers each test,
which means we probably want to hooking into the setUpBeforeClass
and tearDownAfterClass static functions (this means Laravel will
not be booted).

Allow the server host to be set.

When Exceptions occur, it is very easy to end up with a lot of Zombie/
loose processes. Perhaps we can instead only boot up the server in the
main process, instead of forking the process to start the server.