Allow application to be tweaked for a single test
Add new method tweakApplication which will applies a closure to the test
class' application and the server application (for example to tweak the
config for one test). Add another to remove the tweaks, so we don't
pollute other tests.

Simplify the CanServeSiteTrait to only store a single server, which
makes it a lot clearer when applying tweaks. It's unlikely we'll need to
spool up more than one test server.